#190 (start)
* Fixed spectator button refreshing the interface twice when pressing it
* Removed `MU` and `UM` packet, removed mute list (deprecated)
* Removed `IL` packet (deprecated) (ip packet list)
* Moved character specific function its own source file
