+++ dronline-client.pro
@@ -99,6 +99,7 @@ SOURCES += \
   src/charselect.cpp \
   src/commondefs.cpp \
   src/courtroom.cpp \
+  src/courtroom_character.cpp \
   src/courtroom_widgets.cpp \
   src/datatypes.cpp \
   src/debug_functions.cpp \
